+++ inst/NEWS.md
@@ -3,6 +3,8 @@
 
 * Reimplement `parse_iso_8601()` with vectorized code, for speed (#9).
 
+* Fix `parse_date()` and `parse_iso_8601()` for zero-length input (#20).
+
 # Parsedate 1.1.3
 
 * Update URLs in the README